git learning
List a few version control systems 
Github
GitLab
PerForce
What is a repository? 
A repository is a central file storage location. It is used by version control systems to store multiple versions of files.It also tracks the changes made to the file.
Write the sequence of commands that will result in a change to ‘foobar.txt’ to git 
Git commit-m ”change foobar”
Git push
What are the types of merge available with git? Explain them 
Git provides various methods to merge different commits into a base commit. These methods are called Merge Strategies In fast forward, history is just one straight line. When you create a branch, make some commits in that branch, the time you’re ready to merge, there is no new merge on the master. A three-way merge involves three snapshots. Two are the ones that are involved in a two-way merge, and the third one is the base file or the common ancestor with which these two files will be compared.
Create an account on github if you don’t have one already 
https://github.com/snehadubey1
Make a repository called ‘git-learning’, with a single README.md file. 

Put in your answers for the above questions, along with the certificate link and push it to the repository 
In your submission put a link to the repo.